requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
idna==3.7
pycparser==2.22
python-dateutil==2.9.0.post0
python-whois==0.9.4
requests==2.32.3
six==1.16.0
soupsieve==2.5
urllib3==2.2.2
